Add generic bloom filter index creation support in ingestion spec schema definition to support both real time and batch segments
Summary:
Add generic bloom filter index creation support in ingestion spec schema definition to support both real time and batch segments

Added a flag `createBloomFilterIndex` which defaults to false and can be optionally set to true for any String dimension in the ingestion schema to create bloom filter indexes for a dimension which can be used later by broker hosts to prune segments. The segment pruning logic in broker process will now look for both filters on dimensions having bloom filter indexes created in addition to filters on current partition dimensions that are used by some shard specs like HashBasedShardSpec, SingleDimensionShardSpec and BloomFilterNamedShardSpec. The bloom filter index can be enabled regardless of what shard spec is in use and regardless of whether a segment is created by batch or real time ingestion.

```
"dimensionsSpec": {
    "dimensions": [
            {"name": "partner_id", "type":"string", "createBloomFilterIndex": true},
            "eventtype",
            "app",
            {"name": "root_pin_id", "type":"string", "createBloomFilterIndex": true},
            "pin_id",
            "contenttype",
            "pinformat"
      ]
 }
```

Test Plan: Unit test and integration test

Reviewers: O1139 Druid, jgu, itallam

Reviewed By: O1139 Druid, jgu, itallam

Subscribers: jenkins, mleonard, shawncao, #realtime-analytics

Differential Revision: https://phabricator.pinadmin.com/D747062

+++ server/src/main/java/org/apache/druid/segment/realtime/appenderator/AppenderatorImpl.java
@@ -30,7 +30,6 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
-import com.google.common.io.Files;
 import com.google.common.primitives.Ints;
 import com.google.common.util.concurrent.FutureCallback;
 import com.google.common.util.concurrent.Futures;
@@ -44,6 +43,7 @@ import org.apache.druid.client.cache.CachePopulatorStats;
 import org.apache.druid.common.guava.ThreadRenamingCallable;
 import org.apache.druid.data.input.Committer;
 import org.apache.druid.data.input.InputRow;
+import org.apache.druid.data.input.impl.DimensionsSpec;
 import org.apache.druid.java.util.common.DateTimes;
 import org.apache.druid.java.util.common.IAE;
 import org.apache.druid.java.util.common.ISE;
@@ -65,6 +65,7 @@ import org.apache.druid.segment.IndexMerger;
 import org.apache.druid.segment.QueryableIndex;
 import org.apache.druid.segment.QueryableIndexSegment;
 import org.apache.druid.segment.Segment;
+import org.apache.druid.segment.column.ColumnHolder;
 import org.apache.druid.segment.incremental.IncrementalIndexAddResult;
 import org.apache.druid.segment.incremental.IndexSizeExceededException;
 import org.apache.druid.segment.indexing.DataSchema;
@@ -76,7 +77,6 @@ import org.apache.druid.segment.realtime.plumber.Sink;
 import org.apache.druid.server.coordination.DataSegmentAnnouncer;
 import org.apache.druid.timeline.DataSegment;
 import org.apache.druid.timeline.NamespacedVersionedIntervalTimeline;
-import org.apache.druid.timeline.partition.BloomFilterStreamFanOutHashBasedNumberedShardSpec;
 import org.joda.time.Interval;
 
 import javax.annotation.Nullable;
@@ -320,12 +320,6 @@ public class AppenderatorImpl implements Appenderator
     metrics.setRowsInMemory(rowsCurrentlyInMemory.addAndGet(numAddedRows));
     metrics.setBytesInMemory(bytesCurrentlyInMemory.addAndGet(bytesInMemoryAfterAdd - bytesInMemoryBeforeAdd));
     totalRows.addAndGet(numAddedRows);
-
-    if (identifier.getShardSpec() instanceof BloomFilterStreamFanOutHashBasedNumberedShardSpec && numAddedRows > 0) {
-      final long start = System.currentTimeMillis();
-      ((BloomFilterStreamFanOutHashBasedNumberedShardSpec) identifier.getShardSpec()).updateBloomFilter(row);
-      metrics.setUpdateBloomFilterMillis(System.currentTimeMillis() - start);
-    }
 
     PersistType persistType = PersistType.NONE;
     List<String> persistReasons = new ArrayList<>();
@@ -818,8 +812,6 @@ public class AppenderatorImpl implements Appenderator
 
     // Use a descriptor file to indicate that pushing has completed.
     final File persistDir = computePersistDir(identifier);
-    final File mergedTarget = new File(persistDir, "merged");
-    final File mergedSupplimentalIndexTarget = new File(persistDir, "merged_supplimental_index");
     final File descriptorFile = computeDescriptorFile(identifier);
 
     // Sanity checks
@@ -851,68 +843,61 @@ public class AppenderatorImpl implements Appenderator
 
       log.info("Pushing merged index for segment[%s].", identifier);
 
-      removeDirectory(mergedTarget);
-      removeDirectory(mergedSupplimentalIndexTarget);
-
-      if (mergedTarget.exists()) {
-        throw new ISE("Merged target[%s] exists after removing?!", mergedTarget);
+      final File mergedIndexTarget = new File(persistDir, "merged");
+      removeDirectory(mergedIndexTarget);
+      if (mergedIndexTarget.exists()) {
+        throw new ISE("Merged target[%s] exists after removing?!", mergedIndexTarget);
       }
 
-      if (mergedSupplimentalIndexTarget.exists()) {
-        throw new ISE("Merged supplimental index target[%s] exists after removing?!",
-                      mergedSupplimentalIndexTarget);
-      }
-
-      final File mergedFile;
+      final File mergedIndexFile;
+      final File mergedSupplimentalIndexFile;
       List<QueryableIndex> indexes = new ArrayList<>();
       Closer closer = Closer.create();
       try {
+        DimensionsSpec dimensionsSpec = schema.getParser() == null ? null : schema.getParser()
+                                                                                  .getParseSpec()
+                                                                                  .getDimensionsSpec();
+        boolean hasSupplimentalIndex = IndexMerger.hasSupplimentalIndex(dimensionsSpec);
+        final File mergedSupplimentalIndexTarget = hasSupplimentalIndex ?
+                                                   new File(persistDir, "merged_supplimental_index") : null;
+        if (hasSupplimentalIndex) {
+          removeDirectory(mergedSupplimentalIndexTarget);
+          if (mergedSupplimentalIndexTarget.exists()) {
+            throw new ISE(
+                "Merged supplimental index target[%s] exists after removing?!",
+                mergedSupplimentalIndexTarget
+            );
+          }
+        }
+
+        List<String> dimensionNamesHasBloomFilterIndexes = IndexMerger.getDimensionNamesHasBloomFilterIndexes(
+            dimensionsSpec);
         for (FireHydrant fireHydrant : sink) {
           Pair<Segment, Closeable> segmentAndCloseable = fireHydrant.getAndIncrementSegment();
           final QueryableIndex queryableIndex = segmentAndCloseable.lhs.asQueryableIndex();
           log.info("Adding hydrant[%s]", fireHydrant);
+          IndexMerger.setHasBloomFilterIndexesInColumnCapabilities(
+              dimensionNamesHasBloomFilterIndexes,
+              dimension -> {
+                ColumnHolder columnHolder = queryableIndex.getColumnHolder(dimension);
+                return columnHolder == null ? null : columnHolder.getCapabilities();
+              }
+          );
           indexes.add(queryableIndex);
           closer.register(segmentAndCloseable.rhs);
         }
 
-        mergedFile = indexMerger.mergeQueryableIndex(
+        Pair<File, File> p = indexMerger.mergeQueryableIndex(
             indexes,
             schema.getGranularitySpec().isRollup(),
             schema.getAggregators(),
-            mergedTarget,
+            mergedIndexTarget,
+            mergedSupplimentalIndexTarget,
             tuningConfig.getIndexSpec(),
             tuningConfig.getSegmentWriteOutMediumFactory()
         );
-
-        // Upload supplimental indexes
-        if (identifier.getShardSpec() instanceof BloomFilterStreamFanOutHashBasedNumberedShardSpec) {
-          BloomFilterStreamFanOutHashBasedNumberedShardSpec s =
-              ((BloomFilterStreamFanOutHashBasedNumberedShardSpec) identifier.getShardSpec());
-          FileUtils.forceMkdir(mergedSupplimentalIndexTarget);
-
-          s.completeBloomFilter();
-          byte[] serializedBloomFilter = s.serializeBloomFilter();
-          if (serializedBloomFilter != null) {
-            final File partitionDimensionsBloomFilterDir = new File(
-                mergedSupplimentalIndexTarget,
-                BloomFilterStreamFanOutHashBasedNumberedShardSpec.BLOOM_FILTER_DIR
-            );
-            FileUtils.forceMkdir(partitionDimensionsBloomFilterDir);
-            // Write version
-            Files.asByteSink(new File(
-                partitionDimensionsBloomFilterDir,
-                BloomFilterStreamFanOutHashBasedNumberedShardSpec.BLOOM_FILTER_VERSION_FILE
-            )).write(
-                Ints.toByteArray(BloomFilterStreamFanOutHashBasedNumberedShardSpec.CURRENT_VERSION_ID));
-            // Write serialized bloom filter
-            Files.asByteSink(new File(
-                partitionDimensionsBloomFilterDir,
-                BloomFilterStreamFanOutHashBasedNumberedShardSpec.BLOOM_FILTER_BIN_FILE
-            )).write(serializedBloomFilter);
-          } else {
-            log.warn("Empty serialized bytes for bloom filter for segment[%s], skip uploading", identifier);
-          }
-        }
+        mergedIndexFile = p.lhs;
+        mergedSupplimentalIndexFile = p.rhs;
       }
       catch (Throwable t) {
         throw closer.rethrow(t);
@@ -927,8 +912,8 @@ public class AppenderatorImpl implements Appenderator
           // Kafka indexing task, pushers must use unique file paths in deep storage in order to maintain exactly-once
           // semantics.
           () -> dataSegmentPusher.push(
-              mergedFile,
-              mergedSupplimentalIndexTarget,
+              mergedIndexFile,
+              mergedSupplimentalIndexFile,
               sink.getSegment()
                   .withDimensions(IndexMerger.getMergedDimensionsFromQueryableIndexes(indexes)),
               useUniquePath
